Refactor app architecture and align with roadmap
- Frontend: Refactored `App.jsx` to implement routing (Login, Signup, Dashboard) and `ProtectedRoute`. Moved dashboard logic to `src/ui/Dashboard.jsx`.
- Dependencies: Downgraded `react-router-dom` to v6, added `@react-oauth/google`.
- Security: Added security overrides for transitive dependencies.
- Backend: Updated `server/auth.js` to use async bcrypt and handle Google Auth (null passwords).
- NLP: Fixed `src/nlp/cd-detector.js` logic (manual dot product, inverted sensitivity thresholds) and `src/nlp/reframe-engine.js` env vars.
- Testing: Fixed `src/nlp/cd-detector.test.js` to align with logic changes.
- Config: Updated `.gitignore` to exclude log files and build directory.
