+++ package.json
@@ -27,9 +27,10 @@
     "jsonwebtoken": "^9.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^7.11.0",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "^5.0.1",
-    "styled-components": "^6.1.1"
+    "styled-components": "^6.1.1",
+    "@react-oauth/google": "^0.12.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -59,6 +60,10 @@
   "overrides": {
     "nth-check": "^2.0.1",
     "postcss": "^8.4.31",
-    "webpack-dev-server": "^5.2.0"
+    "@isaacs/brace-expansion": "^5.0.1",
+    "@modelcontextprotocol/sdk": "^1.26.0",
+    "fast-xml-parser": "^5.3.4",
+    "hono": "^4.11.7",
+    "tar": "^7.5.7"
   }
 }
