+++ src/nlp/cd-detector.test.js
@@ -21,10 +21,8 @@ jest.mock('@tensorflow/tfjs', () => {
   return {
     ...originalTf,
     mean: jest.fn((tensor) => mockTensor), // Always return mockTensor
-    metrics: {
-      ...originalTf.metrics,
-      cosineDistance: jest.fn((vec1, vec2) => createMockSqueezedTensor(0.9)), // Default high similarity
-    },
+    mul: jest.fn((a, b) => mockTensor),
+    sum: jest.fn((tensor) => createMockSqueezedTensor(0.9)),
     tensor2d: jest.fn(() => mockTensor),
     equal: jest.fn(() => ({
       all: jest.fn(() => ({
@@ -55,7 +53,7 @@ describe('detectCDs', () => {
 
   test('should return an empty array if no cognitive distortions are detected', async () => {
     // Simulate low similarity for all CDs
-    tf.metrics.cosineDistance.mockReturnValue(createMockSqueezedTensor(0.1));
+    tf.sum.mockReturnValue(createMockSqueezedTensor(0.1));
     const text = 'This is a neutral sentence.';
     const result = await detectCDs(text);
     expect(result).toEqual([]);
@@ -64,15 +62,8 @@ describe('detectCDs', () => {
   });
 
   test('should detect "All-or-Nothing" distortion', async () => {
-    // Mock cosineDistance to return high similarity for the relevant CD type
-    // This requires knowing the order of CD types being checked in detectCDs
-    // For simplicity, we'll make all cosine distances high enough to trigger detection
-    // and then filter based on the expected outcome.
-    tf.metrics.cosineDistance.mockImplementation((vec1, vec2) => {
-      // In a real scenario, you might inspect vec1/vec2 to determine which CD is being checked
-      // For now, we'll just return a high similarity for all.
-      return createMockSqueezedTensor(0.95);
-    });
+    // Mock sum to return high similarity
+    tf.sum.mockReturnValue(createMockSqueezedTensor(0.95));
 
     const text = 'I always fail at everything.';
     const result = await detectCDs(text, 'high');
